// src/utils/enums.ts
export enum RequestType {
  GET = "get",
  POST = "post",
  DELETE = "delete",
}

export enum TransactionsCategories {
  HOUSING = 'Housing',
  UTILITIES = 'Utilities',
  TRANSPORTATION = 'Transportation',
  FOOD_AND_DRINKS = 'Food and drinks',
  HEALTH = 'Health',
  ENTERTAINMENT = 'Entertainment',
  PERSONAL_CARE = 'Personal Care',
  DEBT_REPAYMENT = 'Debt Repayment',
  EDUCATION = 'Education',
  TAXES = 'Taxes',
  MISCELLANEOUS = 'Miscellaneous',
  OTHERS = 'Others',
}